refactor(workbench): organize learning code under Discover
