+++ packages/workbench-app/src/lib/app/discover/components/DiscoverNewsArtwork.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import type { DiscoverNewsArtwork } from "../discover-catalog.js";
+import type { DiscoverNewsArtwork } from "../catalog.js";
 
 type Props = {
   artwork: DiscoverNewsArtwork;
