+++ packages/workbench-app/src/lib/app/discover/guides/index.ts
@@ -1,0 +1,2 @@
+export { default as GuideOverlayHost } from "./components/GuideOverlayHost.svelte";
+export { guideState, markGuideCompleted, startGuide } from "./state.svelte.js";
