+++ packages/workbench-app/src/lib/app/discover/guides/setup-content.ts
