+++ packages/workbench-app/src/lib/app/discover/guides/tour-readiness.ts
