+++ packages/workbench-app/src/lib/app/discover/index.ts
@@ -1,0 +1,6 @@
+export { default as DiscoverStartupHost } from "./DiscoverStartupHost.svelte";
+export { default as DiscoverView } from "./components/DiscoverView.svelte";
+export type { DiscoverEditorialAction } from "./catalog.js";
+export { discoverTitlebarCount, discoverUnseenCount } from "./state.svelte.js";
+export { discoverSections, markDiscoverSeen } from "./state.svelte.js";
+export { openDiscoverPane } from "./tabs.svelte.js";
