+++ packages/workbench-app/src/lib/app/discover/tabs.svelte.ts
