--- packages/workbench-app/src/lib/app/onboarding/index.ts
@@ -1,7 +1,0 @@
-export { default as OnboardingHost } from "./components/OnboardingHost.svelte";
-export {
-  discoverTitlebarCount,
-  discoverUnseenCount,
-} from "./discover-state.svelte.js";
-export { openDiscoverPane } from "./discover-tabs.svelte.js";
-export { guideState } from "./guide-state.svelte.js";
